+++ src/services/article/actions.ts
@@ -1,5 +1,6 @@
 import { db } from "@/lib/db";
 import { Article, PublishedArticle } from "@/types/article";
+import { id } from "@instantdb/react";
 
 export function isPublished(article: Article): article is PublishedArticle {
   return article.publishedAt != null;
@@ -11,6 +12,22 @@ export function sortByPublishedAt(articles: Article[]): Article[] {
     const bTime = b.publishedAt ?? 0;
     return bTime - aTime;
   });
+}
+
+export async function createArticle(): Promise<string> {
+  const newId = id();
+  await db.transact(
+    db.tx.articles[newId].create({
+      draftTitle: "",
+      title: "",
+      content: "",
+      draftContent: "",
+      status: "draft",
+      createdAt: Date.now(),
+      updatedAt: Date.now(),
+    }),
+  );
+  return newId;
 }
 
 export async function publishArticle(
